Add EOF command
With this in place, I think it's doable to start integrating this into
the LC-3 OS. Will need to add the following traps (syscalls):

fopen -> if R0 is a pointer to a string of the form /dev/stX, then add
         tape drive X to the fd table. R1 can hold a one-bit read/write
         mode (R=0, W=1). If it's already in the table, exception thrown.
         if X >= 64, exception thrown. Initially rewind on fopen
fputc -> write the char in R0[7:0] to the fd R1. if file open R, exception
fgetc -> read the next char from fd R1 into R0[7:0], or put -1 in R0 if EOF.
         if file open W, exception
ftrunc -> truncate file to end at current position (SET_EOF command)
frewind -> seek to the beginning of the file for fd R1
fclose -> close fd R1 to the fd R1. practically, the OS will just rewind
          and remove from the fd table

For all these, exceptions are handled through the standard LC-3
exception mechanism. This avoids needing to check status codes